[CHERRY-PICK] kernel: signal: Use appropriate user pointer conversion
Cherry pick "4f7d74202908 kernel: signal: Use appropriate user pointer conversion":
| commit 4f7d74202908ff7ef9ee3ca317a86bcb6ba4e134
| Author: Kevin Brodsky <kevin.brodsky@arm.com>
| Date:   Thu Feb 2 12:51:08 2023 +0000
|
|     kernel: signal: Use appropriate user pointer conversion
|
|     si_call_addr is typed as a user pointer, but it really is just an
|     address. Use as_user_ptr() to store the address as a user pointer,
|     unmodified.
|

+++ include/linux/sched/signal.h
@@ -570,7 +570,7 @@ static inline int kill_cad_pid(int sig, int priv)
 #define SEND_SIG_NOINFO ((struct kernel_siginfo *) 0)
 #define SEND_SIG_PRIV	((struct kernel_siginfo *) 1)
 
-static inline int __on_sig_stack(unsigned long sp)
+static inline int __on_sig_stack(uintptr_t sp)
 {
 #ifdef CONFIG_STACK_GROWSUP
 	return sp >= current->sas_ss_sp &&
